fix: post-review corrections for parquet/time types PR
- Add TimeColumn to ComparableOps<&Column> (gt/ge/lt/le/equal/between
  were silently returning None for Time vs Time column comparisons)
- Add ComparableOps<&TimeColumn> for TimeColumn impl
- Fix col-vs-col error messages for Date/DateTime/Time: use human-readable
  strings via i32_to_naive_date / i64_to_naive_datetime / i64_to_naive_time
  instead of raw internal integers
- Fix verdict-py compile error: TimeColumn closure &i32 -> &i64
- Fix parse_usize_array error message: "max must be a string" -> "max must be an integer"
- Rename num_seconds/lower_num_seconds/upper_num_seconds -> micros in
  ComparableOps<&NaiveTime> for TimeColumn (unit is microseconds, not seconds)
- Add ColBuilder::type_name() for readable TypeMismatch error messages
- Add TODO explaining silent-None behavior in ComparableOps<&str> for temporal columns

+++ crates/verdict-py/src/lib.rs
@@ -225,7 +225,7 @@ impl PyColumn {
             Column::Bool(col) => ("bool", format_values(&col.0, |v: &bool| v.to_string())),
             Column::Date(col) => ("date", format_values(&col.0, |v: &i32| v.to_string())),
             Column::DateTime(col) => ("datetime", format_values(&col.0, |v: &i64| v.to_string())),
-            Column::Time(col) => ("time", format_values(&col.0, |v: &i32| v.to_string())),
+            Column::Time(col) => ("time", format_values(&col.0, |v: &i64| v.to_string())),
         };
         format!("[{}]: [{}]", dtype, values)
     }
